+++ src/main/java/io/rscnt/repo/AlbumRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import io.rscnt.model.Album;
 
-public interface AlbumRepo extends JpaRepository<Album, Long> {
+public interface AlbumRepo extends JpaRepository<Album, Integer> {
 
 	@Query("select a from Album a where LOWER(a.nombre) = LOWER(:albumNombre)")
 	public Album findByNombre(@Param("albumNombre") String albumNombre);
